Remove `network.zero` placeholder.
This was a relic from years ago, and putting sites in purgatory is a confusing and bad experience. In this commit, we remove Network Zero all together, and maintain the domain & path when moving sites between networks.

If a site needs to be edited, that can be done manually afterwards.

This also coincides with previously established workflows across aliased domains and complex networks, where primary sites and sub-sites may not be part of the same URL hierarchy. (We do this for profiles.wordpress.org and codex.bbpress.org currently.)

Fixes #63.