Admin: audit usages of FILTER_FORCE_ARRAY. (#172)
This constant is a FLAG, not a filter type. This means it must be used in conjunction with another type of filter, and cannot be used by itself.

This commit adds the FILTER_SANITIZE_NUMBER_INT filter type to these usages, ensuring that their values will be treated as integers.

Fixes #164.